restructure api docs
