+++ main/business/search_business.yml
@@ -67,13 +67,13 @@ responses:
       "Date Created": "Sun, 25 Mar 2018 10:43:31 GMT", 
       "Last Modified": "Mon, 26 Mar 2018 10:23:28 GMT",
       "Review count": 3,
-      "Review url":"https://.../api/v1/businesses/1/reviews", 
       "category": "category 1", 
       "description": "business 1 description", 
       "id": 1, 
       "location": "location 1", 
       "name": "business 1", 
-      "user_id": 1
+      "user_id": 1,
+      "total": 3
     }]
   404:
     description: No businesses by that name
